refactor(vision): one color calibration node, unify use
