+++ nectar/CMakeLists.txt
@@ -37,7 +37,6 @@ install(PROGRAMS
   nectar/interface/app.py
   nectar/vision/nodes/aruco_node.py
   nectar/vision/nodes/color_calibration_node.py
-  nectar/vision/nodes/click_color_calibration_node.py
   nectar/vision/camera/calibration/calibration.py
   nectar/vision/nodes/line_detection_node.py
   nectar/vision/nodes/camera_publisher_node.py
